CloudRunFunction/requirements.txt
Flask==3.0.3
gunicorn==22.0.0
werkzeug==3.0.3
google-cloud-bigquery==3.20.1
requests==2.32.3
